requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.6
django-cors-headers==4.7.0
djangorestframework==3.16.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
python-decouple==3.8
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
